# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /login:
    x-swagger-router-controller: auth
    post:
      description: Returns the user
      operationId: login
      parameters:
        - name: body
          in: body
          description: Username
          schema:
            type: object
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/MovieResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /movies/bestrated:
    x-swagger-router-controller: movies
    get:
      description: Returns the list of best rated movies
      operationId: getBestRatedMovies
      parameters:
        - name: page
          in: query
          description: The page to fetch from the movie API
          required: false
          type: integer
        - name: username
          in: query
          description: The username of the logged in user to check his subscriptions
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/MovieResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /subscribe:
    x-swagger-router-controller: subscriptions
    post:
      description: Subscribe to a movie
      operationId: postSubscribe
      parameters:
        - name: body
          in: body
          description: the body
          schema:
            type: object
            properties:
              username:
                type: string
              movie:
                type: integer
      responses:
        "201":
          description: Created
          schema:
            $ref: "#/definitions/MovieResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /unsubscribe:
    x-swagger-router-controller: subscriptions
    post:
      description: Unsubscribe to a movie
      operationId: postUnsubscribe
      parameters:
        - name: body
          in: body
          description: the body
          schema:
            type: object
            properties:
              username:
                type: string
              movie:
                type: integer
      responses:
        "200":
          description: Ok
          schema:
            $ref: "#/definitions/MovieResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  MovieResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string